refactor(ssf): remove old review / subject-filter CLI commands (#111)
