+++ cmd/goSignals/commands.go
@@ -1975,91 +1975,6 @@ type TokenCmd struct {
 	Introspect TokenIntrospectCmd `cmd:"" help:"Introspect a token (RFC7662)"`
 }
 
-// ReviewSubjectFilterCmd is the operator-facing CLI for PRD #97 issue #101: a
-// read-only view of a transmitter stream's locally managed SSF §8.1.3 subject
-// filter. The summary (counts + pending-removal list) is the default; an
-// optional --subject JSON literal opts a point-lookup result into the output.
-// The CLI never dumps the full filter table — the server endpoint is
-// point-lookup + counts by design (ADR-0003).
-type ReviewSubjectFilterCmd struct {
-	Alias   string `arg:"" optional:"" help:"Stream alias to review (defaults to the selected stream)."`
-	Subject string `optional:"" short:"s" help:"SubjectIdentifier JSON for a point lookup, e.g. '{\"format\":\"email\",\"email\":\"alice@example.com\"}'."`
-}
-
-func (r *ReviewSubjectFilterCmd) Run(cli *CLI) error {
-	alias := r.Alias
-	if alias == "" {
-		alias = cli.Data.Selected
-	}
-	stream, server := cli.Data.GetStreamAndServer(alias)
-	if stream == nil {
-		return errors.New("Could not locate locally defined stream alias: " + alias)
-	}
-
-	body := map[string]any{"stream_id": stream.Id}
-	if r.Subject != "" {
-		var subject goSet.SubjectIdentifier
-		if err := json.Unmarshal([]byte(r.Subject), &subject); err != nil {
-			return fmt.Errorf("--subject must be a SubjectIdentifier JSON literal: %w", err)
-		}
-		body["subject"] = subject
-	}
-	reqBody, err := json.Marshal(body)
-	if err != nil {
-		return err
-	}
-
-	reviewUrl, err := url.JoinPath(server.Host, "/subject-filter/review")
-	if err != nil {
-		return err
-	}
-	req, err := http.NewRequest(http.MethodPost, reviewUrl, bytes.NewReader(reqBody))
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Authorization", "Bearer "+server.ClientToken)
-	req.Header.Set("Content-Type", "application/json")
-
-	client := getHttpClient(0)
-	defer client.CloseIdleConnections()
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer httpSupport.HandleRespClose(resp)
-	respBody, _ := io.ReadAll(resp.Body)
-	if resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("review request failed: %s: %s", resp.Status, string(respBody))
-	}
-
-	// Pretty-print the JSON if it parses; otherwise pass it through raw.
-	var pretty bytes.Buffer
-	if jerr := json.Indent(&pretty, respBody, "", "  "); jerr == nil {
-		fmt.Println("Subject filter review for " + alias + ":")
-		fmt.Println(pretty.String())
-		cli.GetOutputWriter().WriteBytes(pretty.Bytes(), true)
-	} else {
-		fmt.Println(string(respBody))
-		cli.GetOutputWriter().WriteBytes(respBody, true)
-	}
-	return nil
-}
-
-type ReviewCmd struct {
-	SubjectFilter ReviewSubjectFilterCmd `cmd:"" aliases:"sf" help:"Review a stream's locally managed SSF subject filter (admin-scoped)."`
-}
-
-// SubjectFilterShowCmd reads the four subject-filtering operator knobs for a
-// stream — defaultSubjects, subject-filter mode, event source, and the SSF §9.3
-// removal grace override — and prints them (PRD #97 issue #102). It hits the
-// existing admin-scoped /subject-filter/review endpoint, which already returns
-// the policy fields alongside the filter-table summary; this command formats
-// only the policy bits for an operator who wants the settings without the
-// counts/pending list.
-type SubjectFilterShowCmd struct {
-	Alias string `arg:"" optional:"" help:"Stream alias to show (defaults to the selected stream)."`
-}
-
 // subjectFilterReviewWire mirrors the wire shape of the admin review endpoint
 // (api_subject_filter_review.go) — just the policy fields the CLI prints. Kept
 // local so the CLI is decoupled from the server's internal types.
@@ -2139,126 +2054,6 @@ func formatSubjectFilterSettings(alias string, wire *subjectFilterReviewWire) st
 	}
 	return fmt.Sprintf("Subject-filter settings for [%s]:\n  stream_id:                     %s\n  default_subjects:              %s\n  mode:                          %s\n  event_source:                  %s\n  subject_removal_grace_seconds: %d\n",
 		alias, wire.StreamId, defaults, mode, es, wire.SubjectRemovalGraceSeconds)
-}
-
-func (s *SubjectFilterShowCmd) Run(cli *CLI) error {
-	alias := s.Alias
-	if alias == "" {
-		alias = cli.Data.Selected
-	}
-	stream, server := cli.Data.GetStreamAndServer(alias)
-	if stream == nil {
-		return errors.New("Could not locate locally defined stream alias: " + alias)
-	}
-	wire, err := fetchSubjectFilterSettings(server, stream.Id)
-	if err != nil {
-		return err
-	}
-	out := formatSubjectFilterSettings(alias, wire)
-	fmt.Print(out)
-	cli.GetOutputWriter().WriteString(out, true)
-	return nil
-}
-
-// SubjectFilterSetCmd writes the four subject-filtering operator knobs through
-// PRD #89's existing stream-update path (PRD #97 issue #102). No new server
-// endpoint is added: the JSON body PUT to /stream is a StreamStateRecord shape
-// — the four knobs ride at the top level alongside the embedded
-// StreamConfiguration — and the server's StreamUpdate treats them as a partial
-// update (empty/zero means "do not change"). Setting `--grace-seconds` on a
-// receiver stream is ignored server-side with a WARN; the post-update display
-// surfaces that ignore by re-reading the persisted value via the review
-// endpoint, so the operator sees what actually landed.
-type SubjectFilterSetCmd struct {
-	Alias           string `arg:"" optional:"" help:"Stream alias to update (defaults to the selected stream)."`
-	DefaultSubjects string `optional:"" default:"" enum:"ALL,NONE," help:"Baseline policy (ALL or NONE). Omit to leave unchanged."`
-	Mode            string `optional:"" default:"" enum:"PASSTHRU,LOCAL,HYBRID," help:"Subject-filter mode for a receiver stream. Omit to leave unchanged."`
-	EventSource     string `optional:"" default:"" enum:"DIRECT,AUDIENCE,EXPLICIT," help:"Event source type for a transmitter stream. Omit to leave unchanged."`
-	GraceSeconds    *int   `optional:"" help:"Per-transmitter-stream removal grace period override in seconds (SSF §9.3). 0 means immediate; omit to leave unchanged."`
-}
-
-func (s *SubjectFilterSetCmd) Run(cli *CLI) error {
-	alias := s.Alias
-	if alias == "" {
-		alias = cli.Data.Selected
-	}
-	stream, server := cli.Data.GetStreamAndServer(alias)
-	if stream == nil {
-		return errors.New("Could not locate locally defined stream alias: " + alias)
-	}
-
-	// Build a partial-update body. The server's StreamUpdate reads it into a
-	// StreamStateRecord and treats empty/zero fields as "no change". The
-	// embedded StreamConfiguration is left untouched — only the operator-knob
-	// fields are populated.
-	body := map[string]any{
-		"stream_id": stream.Id,
-		// Echo the stream_id on the embedded configuration too so the server's
-		// StreamUpdate authorization check finds it via configReq.Id.
-		"id": stream.Id,
-	}
-	if s.DefaultSubjects != "" {
-		body["default_subjects"] = s.DefaultSubjects
-	}
-	if s.Mode != "" {
-		body["subject_filter_mode"] = s.Mode
-	}
-	if s.EventSource != "" {
-		body["event_source"] = map[string]any{"type": s.EventSource}
-	}
-	if s.GraceSeconds != nil {
-		body["subject_removal_grace_seconds"] = *s.GraceSeconds
-	}
-
-	reqBytes, err := json.Marshal(body)
-	if err != nil {
-		return err
-	}
-	req, err := http.NewRequest(http.MethodPut,
-		server.ServerConfiguration.ConfigurationEndpoint+"?stream_id="+stream.Id,
-		bytes.NewReader(reqBytes))
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Authorization", "Bearer "+server.ClientToken)
-	req.Header.Set("Content-Type", "application/json")
-	client := getHttpClient(0)
-	defer client.CloseIdleConnections()
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer httpSupport.HandleRespClose(resp)
-	respBody, _ := io.ReadAll(resp.Body)
-	if resp.StatusCode != http.StatusOK {
-		// PRD #89 / #98 validation failures surface here with the server's
-		// error message — e.g. an invalid LOCAL/HYBRID combination or a
-		// negative grace value.
-		return fmt.Errorf("stream update failed: %s: %s", resp.Status, strings.TrimSpace(string(respBody)))
-	}
-	cli.Data.ResetStreamConfig(alias)
-
-	// Read back the post-update settings so the operator sees what actually
-	// landed. This surfaces the server's WARN/ignore behavior for a grace
-	// override set on a receiver stream — the persisted value comes back as 0.
-	wire, err := fetchSubjectFilterSettings(server, stream.Id)
-	if err != nil {
-		return err
-	}
-	out := formatSubjectFilterSettings(alias, wire)
-	fmt.Print(out)
-	cli.GetOutputWriter().WriteString(out, true)
-	return nil
-}
-
-// SubjectFilterCmd is the operator-facing command group for managing a stream's
-// SSF subject-filtering settings (PRD #97 issue #102): one coherent place to
-// view and change defaultSubjects, mode, event source, and the SSF §9.3
-// removal grace override. The existing `review subject-filter` command remains
-// focused on filter-table state (counts, pending list, point lookup).
-type SubjectFilterCmd struct {
-	Show SubjectFilterShowCmd `cmd:"" help:"Show a stream's subject-filtering settings (defaultSubjects, mode, event source, grace override)."`
-	Set  SubjectFilterSetCmd  `cmd:"" help:"Set a stream's subject-filtering settings via the existing stream-update path."`
 }
 
 // subjectFilterStatusWire mirrors the runtime-state portion of the admin
